Gowheels/requirements.txt
Django==4.2.7
PyMySQL==1.1.0
python-decouple==3.8
Pillow==11.0.0